Cast the result of __builtin_* calls to unsigned types
The GCC/clang builtins listed here return int even though they never
return a negative number.

The __pony_* wrappers of these builtins all return unsigned and most
of the use-cases of the builtins make sense with the unsigned result,
so I think it makes sense to have unsigned return as the default for
all platforms.

This change means we don't have to add many explicit casts in the calls
to __pony_* to fix -Wsign-conversion violations.

+++ src/common/platform.h
@@ -232,13 +232,13 @@ inline int snprintf(char* str, size_t size, const char* format, ...)
  *
  */
 #ifdef PLATFORM_IS_CLANG_OR_GCC
-#  define __pony_popcount(X) __builtin_popcount((X))
-#  define __pony_ffs(X) __builtin_ffs((X))
-#  define __pony_ffsl(X) __builtin_ffsl((X))
-#  define __pony_ctz(X) __builtin_ctz((X))
-#  define __pony_ctzl(X) __builtin_ctzl((X))
-#  define __pony_clz(X) __builtin_clz((X))
-#  define __pony_clzl(X) __builtin_clzl((X))
+#  define __pony_popcount(X) (unsigned int)__builtin_popcount((X))
+#  define __pony_ffs(X) (unsigned int)__builtin_ffs((X))
+#  define __pony_ffsl(X) (unsigned int)__builtin_ffsl((X))
+#  define __pony_ctz(X) (unsigned int)__builtin_ctz((X))
+#  define __pony_ctzl(X) (unsigned int)__builtin_ctzl((X))
+#  define __pony_clz(X) (unsigned int)__builtin_clz((X))
+#  define __pony_clzl(X) (unsigned int)__builtin_clzl((X))
 #else
 #  include <intrin.h>
 #  define __pony_popcount(X) __popcnt((X))
